refactor(MainPage): main 페이지 index 페이지로 수정
