+++ pages/index.tsx
@@ -1,72 +1,50 @@
-import type { NextPage } from 'next'
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
+import styled from '@emotion/styled'
+import PostCard from '..components/PostCard'
 
-const Home: NextPage = () => {
+const PostCardDummy = Array.from({ length: 5 }, (_, idx) => {
+  return {
+    id: idx.toString(),
+    title: '참치마요',
+    imageUrl: 'https://via.placeholder.com/300x150',
+    avatarImageUrl: 'https://via.placeholder.com/50',
+    author: 'Lee',
+    likes: 10,
+    comments: 20
+  }
+})
+
+const Main = () => {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+    <TempLayout>
+      {PostCardDummy.map((cardData) => {
+        return (
+          <PostCard
+            id={cardData.id}
+            key={cardData.id}
+            title={cardData.title}
+            imageUrl={cardData.imageUrl}
+            avatarImageUrl={cardData.avatarImageUrl}
+            author={cardData.author}
+            likes={cardData.likes}
+            comments={cardData.comments}
+          />
+        )
+      })}
+    </TempLayout>
   )
 }
 
-export default Home
+const TempLayout = styled.div`
+  position: relative;
+  left: 20%;
+  width: 37.5rem;
+  height: 81.2rem;
+  border: 1px solid black;
+  overflow-y: scroll;
+
+  ::-webkit-scrollbar {
+    display: none;
+  }
+`
+
+export default Main
